Notifications are now receiving in foreground
- Need to run notifications in background.
- Added new folder to store time based activity screens
