+++ Front_End/app/(root)/(usage-timer)/_layout.tsx
@@ -1,0 +1,12 @@
+import { Stack } from "expo-router";
+
+const TimerLayout = () => {
+  return (
+    <Stack>
+      <Stack.Screen name="usage-limits" options={{ headerShown: false }} />
+      <Stack.Screen name="app-timer" options={{ headerShown: false }} />
+    </Stack>
+  );
+};
+
+export default TimerLayout;
