+++ Front_End/app/(root)/_layout.tsx
@@ -4,6 +4,7 @@ const Layout = () => {
   return (
     <Stack>
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+      <Stack.Screen name="(usage-timer)" options={{ headerShown: false }} />
     </Stack>
   );
