+++ Front_End/app/_layout.tsx
@@ -47,6 +47,7 @@ export default function RootLayout() {
           <Stack.Screen name="index" options={{ headerShown: false }} />
           <Stack.Screen name="(root)" options={{ headerShown: false }} />
           <Stack.Screen name="(auth)" options={{ headerShown: false }} />
+          
         </Stack>
       </ClerkLoaded>
     </ClerkProvider>
