+++ Front_End/package.json
@@ -25,6 +25,7 @@
     "@react-navigation/native": "^7.0.14",
     "expo": "~52.0.31",
     "expo-auth-session": "~6.0.3",
+    "expo-background-fetch": "~13.0.5",
     "expo-blur": "~14.0.3",
     "expo-constants": "~17.0.5",
     "expo-font": "~13.0.3",
@@ -46,14 +47,14 @@
     "react-native": "0.76.7",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-modal": "^13.0.1",
+    "react-native-onesignal": "^5.2.9",
     "react-native-reanimated": "~3.16.1",
     "react-native-safe-area-context": "4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-svg": "^15.11.1",
     "react-native-swiper": "^1.6.0",
     "react-native-web": "~0.19.13",
-    "react-native-webview": "13.12.5",
-    "expo-background-fetch": "~13.0.5"
+    "react-native-webview": "13.12.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
